[TD-884] 가게 조회랑 커맨드 API 분리 및 기타 리팩토링 진행
