+++ threedollar-api-admin-service/src/main/kotlin/com/depromeet/threedollar/api/adminservice/controller/commonservice/faq/AdminFaqController.kt
@@ -15,8 +15,8 @@ import com.depromeet.threedollar.api.adminservice.service.commonservice.faq.dto.
 import com.depromeet.threedollar.api.adminservice.service.commonservice.faq.dto.request.UpdateFaqRequest
 import com.depromeet.threedollar.api.adminservice.service.commonservice.faq.dto.response.FaqAdminResponse
 import com.depromeet.threedollar.api.core.common.dto.ApiResponse
-import com.depromeet.threedollar.api.core.service.commonservice.faq.FaqCategoryMapper
-import com.depromeet.threedollar.api.core.service.commonservice.faq.dto.response.FaqCategoryResponse
+import com.depromeet.threedollar.api.core.mapper.commonservice.faq.FaqCategoryMapper
+import com.depromeet.threedollar.api.core.mapper.commonservice.faq.dto.response.FaqCategoryResponse
 import com.depromeet.threedollar.api.core.service.commonservice.faq.dto.response.FaqResponse
 import com.depromeet.threedollar.common.type.ApplicationType
 import com.depromeet.threedollar.domain.rds.domain.commonservice.faq.FaqCategory
