+++ threedollar-api-boss-service/src/main/kotlin/com/depromeet/threedollar/api/bossservice/controller/LocalTestController.kt
@@ -15,11 +15,11 @@ import com.depromeet.threedollar.api.bossservice.config.interceptor.Auth
 import com.depromeet.threedollar.api.bossservice.config.resolver.BossId
 import com.depromeet.threedollar.api.bossservice.config.session.SessionConstants
 import com.depromeet.threedollar.api.bossservice.service.auth.dto.response.LoginResponse
-import com.depromeet.threedollar.api.bossservice.service.store.BossStoreServiceHelper
 import com.depromeet.threedollar.api.core.common.dto.ApiResponse
 import com.depromeet.threedollar.api.core.service.bossservice.category.BossStoreCategoryServiceHelper
 import com.depromeet.threedollar.api.core.service.bossservice.feedback.BossStoreFeedbackService
 import com.depromeet.threedollar.api.core.service.bossservice.feedback.dto.request.AddBossStoreFeedbackRequest
+import com.depromeet.threedollar.api.core.service.bossservice.store.BossStoreServiceHelper
 import com.depromeet.threedollar.common.exception.model.ConflictException
 import com.depromeet.threedollar.common.exception.model.InternalServerException
 import com.depromeet.threedollar.common.exception.model.NotFoundException
