+++ threedollar-api-boss-service/src/main/kotlin/com/depromeet/threedollar/api/bossservice/controller/faq/BossFaqController.kt
@@ -4,10 +4,10 @@ import javax.validation.Valid
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RestController
 import com.depromeet.threedollar.api.core.common.dto.ApiResponse
-import com.depromeet.threedollar.api.core.service.commonservice.faq.FaqCategoryMapper
+import com.depromeet.threedollar.api.core.mapper.commonservice.faq.FaqCategoryMapper
+import com.depromeet.threedollar.api.core.mapper.commonservice.faq.dto.response.FaqCategoryResponse
 import com.depromeet.threedollar.api.core.service.commonservice.faq.FaqService
 import com.depromeet.threedollar.api.core.service.commonservice.faq.dto.request.RetrieveFaqsRequest
-import com.depromeet.threedollar.api.core.service.commonservice.faq.dto.response.FaqCategoryResponse
 import com.depromeet.threedollar.api.core.service.commonservice.faq.dto.response.FaqResponse
 import com.depromeet.threedollar.common.type.ApplicationType
 import io.swagger.annotations.ApiOperation
