+++ threedollar-api-boss-service/src/main/kotlin/com/depromeet/threedollar/api/bossservice/service/store/BossStoreOpenService.kt
@@ -2,6 +2,7 @@ package com.depromeet.threedollar.api.bossservice.service.store
 
 import java.time.LocalDateTime
 import org.springframework.stereotype.Service
+import com.depromeet.threedollar.api.core.service.bossservice.store.BossStoreServiceHelper
 import com.depromeet.threedollar.common.model.LocationValue
 import com.depromeet.threedollar.domain.mongo.domain.bossservice.store.BossStoreRepository
 import com.depromeet.threedollar.domain.redis.domain.bossservice.store.BossStoreOpenTimeRepository
