+++ threedollar-api-boss-service/src/main/kotlin/com/depromeet/threedollar/api/bossservice/service/store/BossStoreService.kt
@@ -4,23 +4,19 @@ import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import com.depromeet.threedollar.api.bossservice.service.store.dto.request.PatchBossStoreInfoRequest
 import com.depromeet.threedollar.api.bossservice.service.store.dto.request.UpdateBossStoreInfoRequest
-import com.depromeet.threedollar.api.core.service.bossservice.category.BossStoreCategoryService
 import com.depromeet.threedollar.api.core.service.bossservice.category.BossStoreCategoryServiceHelper
-import com.depromeet.threedollar.api.core.service.bossservice.store.dto.response.BossStoreInfoResponse
+import com.depromeet.threedollar.api.core.service.bossservice.store.BossStoreServiceHelper
 import com.depromeet.threedollar.domain.mongo.domain.bossservice.category.BossStoreCategoryRepository
 import com.depromeet.threedollar.domain.mongo.domain.bossservice.store.BossDeletedStore
 import com.depromeet.threedollar.domain.mongo.domain.bossservice.store.BossDeletedStoreRepository
 import com.depromeet.threedollar.domain.mongo.domain.bossservice.store.BossStore
 import com.depromeet.threedollar.domain.mongo.domain.bossservice.store.BossStoreRepository
-import com.depromeet.threedollar.domain.redis.domain.bossservice.store.BossStoreOpenTimeRepository
 
 @Service
 class BossStoreService(
     private val bossStoreRepository: BossStoreRepository,
     private val bossDeleteBossStoreRepository: BossDeletedStoreRepository,
     private val bossStoreCategoryRepository: BossStoreCategoryRepository,
-    private val bossStoreOpenTimeRepository: BossStoreOpenTimeRepository,
-    private val bossStoreCategoryService: BossStoreCategoryService,
 ) {
 
     @Transactional
@@ -79,16 +75,6 @@ class BossStoreService(
             bossDeleteBossStoreRepository.save(BossDeletedStore.of(it))
             bossStoreRepository.delete(it)
         }
-    }
-
-    @Transactional(readOnly = true)
-    fun getMyBossStore(bossId: String): BossStoreInfoResponse {
-        val bossStore = BossStoreServiceHelper.findBossStoreByBossId(bossStoreRepository = bossStoreRepository, bossId = bossId)
-        return BossStoreInfoResponse.of(
-            bossStore = bossStore,
-            categories = bossStoreCategoryService.retrieveBossStoreCategoriesByIds(bossStore.categoriesIds),
-            openStartDateTime = bossStoreOpenTimeRepository.get(bossStore.id)
-        )
     }
 
 }
