+++ threedollar-api-user-service/src/e2e-test/java/com/depromeet/threedollar/api/userservice/controller/faq/FaqMockApiCaller.java
@@ -11,7 +11,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
 
 import com.depromeet.threedollar.api.core.common.dto.ApiResponse;
-import com.depromeet.threedollar.api.core.service.commonservice.faq.dto.response.FaqCategoryResponse;
+import com.depromeet.threedollar.api.core.mapper.commonservice.faq.dto.response.FaqCategoryResponse;
 import com.depromeet.threedollar.api.core.service.commonservice.faq.dto.response.FaqResponse;
 import com.depromeet.threedollar.api.userservice.MockMvcHelper;
 import com.depromeet.threedollar.domain.rds.domain.commonservice.faq.FaqCategory;
