+++ threedollar-api-user-service/src/main/java/com/depromeet/threedollar/api/userservice/service/visit/VisitHistoryService.java
@@ -1,5 +1,7 @@
 package com.depromeet.threedollar.api.userservice.service.visit;
 
+import static com.depromeet.threedollar.common.exception.type.ErrorCode.CONFLICT_VISIT_HISTORY;
+
 import java.time.LocalDate;
 import java.util.List;
 
@@ -10,6 +12,7 @@ import com.depromeet.threedollar.api.userservice.service.store.StoreServiceHelpe
 import com.depromeet.threedollar.api.userservice.service.visit.dto.request.AddVisitHistoryRequest;
 import com.depromeet.threedollar.api.userservice.service.visit.dto.request.RetrieveMyVisitHistoriesRequest;
 import com.depromeet.threedollar.api.userservice.service.visit.dto.response.VisitHistoriesCursorResponse;
+import com.depromeet.threedollar.common.exception.model.ConflictException;
 import com.depromeet.threedollar.domain.rds.core.support.CursorPagingSupporter;
 import com.depromeet.threedollar.domain.rds.domain.userservice.store.Store;
 import com.depromeet.threedollar.domain.rds.domain.userservice.store.StoreRepository;
@@ -28,8 +31,14 @@ public class VisitHistoryService {
     @Transactional
     public Long addVisitHistory(AddVisitHistoryRequest request, Long userId, LocalDate dateOfVisit) {
         Store store = StoreServiceHelper.findStoreById(storeRepository, request.getStoreId());
-        VisitHistoryServiceHelper.validateNotVisitedToday(visitHistoryRepository, request.getStoreId(), userId, dateOfVisit);
+        validateNotVisitedToday(request.getStoreId(), userId, dateOfVisit);
         return visitHistoryRepository.save(request.toEntity(store, userId, dateOfVisit)).getId();
+    }
+
+    private void validateNotVisitedToday(Long storeId, Long userId, LocalDate date) {
+        if (visitHistoryRepository.existsByStoreIdAndUserIdAndDateOfVisit(storeId, userId, date)) {
+            throw new ConflictException(String.format("유저(%s)는 해당 날짜(%s)에 유저 가게(%s)를 이미 방문 인증하였습니다", userId, date, storeId), CONFLICT_VISIT_HISTORY);
+        }
     }
 
     @Transactional(readOnly = true)
